+++ src/QueryEditor.tsx
@@ -1,29 +1,27 @@
 import { QueryEditorProps, SelectableValue } from '@grafana/data';
-import { LegacyForms, InlineFormLabel } from '@grafana/ui';
+import { AsyncSelect, InlineFormLabel } from '@grafana/ui';
 import React, { PureComponent } from 'react';
+import debounce from 'debounce-promise';
 
 import { SdsDataSource } from './DataSource';
 import { SdsDataSourceOptions, SdsQuery } from './types';
 
-const { AsyncSelect } = LegacyForms;
-
 type Props = QueryEditorProps<SdsDataSource, SdsQuery, SdsDataSourceOptions>;
 
 export class QueryEditor extends PureComponent<Props> {
-  streams: Array<SelectableValue<string>> = [];
-
   constructor(props: Props) {
     super(props);
-  }
-
-  async queryStreamsAsync(value: string) {
-    return await this.props.datasource.getStreams(value);
   }
 
   onSelectedStream = (value: SelectableValue<string>) => {
     const { onChange, query } = this.props;
     onChange({ ...query, streamId: value.value || '', streamName: value.label || '' });
   };
+
+  debouncedGetStreams = debounce((inputvalue: string) => {
+    return this.props.datasource.getStreams(inputvalue)}, 
+    1000
+  );
 
   render() {
     const query = this.props.query;
@@ -34,16 +32,14 @@ export class QueryEditor extends PureComponent<Props> {
         <InlineFormLabel width={8}>Stream</InlineFormLabel>
         <AsyncSelect
           defaultOptions={true}
-          width={20}
-          loadOptions={(inputvalue) => this.queryStreamsAsync(inputvalue)}
-          defaultValue={query.streamId}
+          width={50}
+          loadOptions={(inputvalue) => this.debouncedGetStreams(inputvalue)}
           value={selectStream}
           onChange={(inputvalue) => this.onSelectedStream(inputvalue)}
           placeholder="Select Stream"
-          loadingMessage={() => 'Loading streams...'}
-          noOptionsMessage={() => 'No streams found'}
+          loadingMessage={'Loading streams...'}
+          noOptionsMessage={'No streams found'}
         />
-        <LegacyForms.Input value={query.streamId || ''} readOnly={true} hidden />
       </div>
     );
   }